modified SecContent.js with props moved to App.js
